Refactored some stuff;
