+++ src/main/java/com/theonlytails/ruby/init/ToolsRegistry.java
@@ -13,7 +13,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 public class ToolsRegistry {
     public static final DeferredRegister<Item> ITEMS =
             DeferredRegister.create(ForgeRegistries.ITEMS, TheOnlyTails.MOD_ID);
-  
+
     public static final RegistryObject<PickaxeItem> RUBY_PICKAXE =
             // A custom pickaxe made out of ruby
             ItemsRegistry.ITEMS.register("ruby_pickaxe", () ->
