+++ src/main/java/com/theonlytails/ruby/world/gen/ModEntitySpawns.java
@@ -1,7 +1,7 @@
 package com.theonlytails.ruby.world.gen;
 
 import com.theonlytails.ruby.TheOnlyTails;
-import com.theonlytails.ruby.init.RubyEntityTypes;
+import com.theonlytails.ruby.init.EntityTypesRegistry;
 import net.minecraft.entity.EntityClassification;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.Biomes;
@@ -24,7 +24,7 @@ public class ModEntitySpawns {
 
                 if (biome != Biomes.WOODED_BADLANDS_PLATEAU && biome != Biomes.SNOWY_TUNDRA) {
                     biome.getSpawns(EntityClassification.CREATURE)
-                            .add(new Biome.SpawnListEntry(RubyEntityTypes.RUBY_SHEEP.get(),
+                            .add(new Biome.SpawnListEntry(EntityTypesRegistry.RUBY_SHEEP.get(),
                                     12, 2, 3));
                 }
             }
